use/install.rst: use `sudo`
With help from \#supybot I understood that telling user to add `sudo` in
front of every command isn't very user-friendly.

If someone has root and is logged in as root, they know to remove the
`sudo` part and if they don't, but they have `sudo` installed,
everything works directly and `sudo` doesn't even ask for password.